refactor: useAuth를 auth 폴더로 이동
* useLogin의 ui 로직 제거
