+++ src/components/Layout/Sidebar.tsx
@@ -10,7 +10,7 @@ import {
 import { useNavigate } from "react-router-dom";
 import { v4 as uuidv4 } from "uuid";
 
-import useAuth from "@/hooks/useAuth";
+import useAuth from "@/features/auth/hooks/useAuth";
 
 import Avatar from "../Avatar";
 import Button from "../Button";
